Fix browser test failures
SharedKeyCredential is only available on Node.js runtime. This change moves all
connection string tests to node tests.
