+++ sdk/storage/storage-blob/src/utils/utils.common.ts
@@ -71,6 +71,8 @@ export function escapeURLPath(url: string): string {
 }
 
 /**
+ * ONLY AVAILABLE IN NODE.JS RUNTIME.
+ *
  * Extracts the parts of an Azure Storage account connection string.
  *
  * @export
