+++ sdk/storage/storage-blob/test/blobserviceclient.spec.ts
@@ -1,15 +1,8 @@
 import * as assert from "assert";
 
 import * as dotenv from "dotenv";
-import { newPipeline, SharedKeyCredential } from "../src";
 import { BlobServiceClient } from "../src/BlobServiceClient";
-import {
-  getAlternateBSU,
-  getBSU,
-  getConnectionStringFromEnvironment,
-  getUniqueName,
-  wait
-} from "./utils";
+import { getAlternateBSU, getBSU, getUniqueName, wait } from "./utils";
 dotenv.config({ path: "../.env" });
 
 describe("BlobServiceClient", () => {
@@ -288,61 +281,5 @@ describe("BlobServiceClient", () => {
     } catch (error) {
       assert.ok((error.statusCode as number) === 404);
     }
-  });
-
-  it("can be created with a url and a credential", async () => {
-    const serviceClient = getBSU();
-    const factories = serviceClient.pipeline.factories;
-    const credential = factories[factories.length - 1] as SharedKeyCredential;
-    const newClient = new BlobServiceClient(serviceClient.url, credential);
-
-    const result = await newClient.getProperties();
-
-    assert.ok(typeof result.requestId);
-    assert.ok(result.requestId!.length > 0);
-    assert.ok(typeof result.version);
-    assert.ok(result.version!.length > 0);
-  });
-
-  it("can be created with a url and a credential and an option bag", async () => {
-    const serviceClient = getBSU();
-    const factories = serviceClient.pipeline.factories;
-    const credential = factories[factories.length - 1] as SharedKeyCredential;
-    const newClient = new BlobServiceClient(serviceClient.url, credential, {
-      retryOptions: {
-        maxTries: 5
-      }
-    });
-
-    const result = await newClient.getProperties();
-
-    assert.ok(typeof result.requestId);
-    assert.ok(result.requestId!.length > 0);
-    assert.ok(typeof result.version);
-    assert.ok(result.version!.length > 0);
-  });
-
-  it("can be created with a url and a pipeline", async () => {
-    const serviceClient = getBSU();
-    const factories = serviceClient.pipeline.factories;
-    const credential = factories[factories.length - 1] as SharedKeyCredential;
-    const pipeline = newPipeline(credential);
-    const newClient = new BlobServiceClient(serviceClient.url, pipeline);
-
-    const result = await newClient.getProperties();
-
-    assert.ok(typeof result.requestId);
-    assert.ok(result.requestId!.length > 0);
-    assert.ok(typeof result.version);
-    assert.ok(result.version!.length > 0);
-  });
-
-  it("can be created from a connection string", async () => {
-    const newClient = BlobServiceClient.fromConnectionString(getConnectionStringFromEnvironment());
-
-    const result = await newClient.getProperties();
-
-    assert.ok(typeof result.requestId);
-    assert.ok(result.requestId!.length > 0);
   });
 });
