+++ sdk/storage/storage-blob/test/blockblobclient.spec.ts
@@ -1,15 +1,7 @@
 import * as assert from "assert";
 
 import * as dotenv from "dotenv";
-import { newPipeline, SharedKeyCredential } from "../src";
-import { BlockBlobClient } from "../src/BlockBlobClient";
-import {
-  base64encode,
-  bodyToString,
-  getBSU,
-  getConnectionStringFromEnvironment,
-  getUniqueName
-} from "./utils";
+import { base64encode, bodyToString, getBSU, getUniqueName } from "./utils";
 dotenv.config({ path: "../.env" });
 
 describe("BlockBlobClient", () => {
@@ -195,84 +187,5 @@ describe("BlockBlobClient", () => {
     assert.equal(listResponse.uncommittedBlocks!.length, 0);
     assert.equal(listResponse.committedBlocks![0].name, base64encode("2"));
     assert.equal(listResponse.committedBlocks![0].size, body.length);
-  });
-
-  it("can be created with a url and a credential", async () => {
-    const factories = blockBlobClient.pipeline.factories;
-    const credential = factories[factories.length - 1] as SharedKeyCredential;
-    const newClient = new BlockBlobClient(blockBlobClient.url, credential);
-
-    const body: string = getUniqueName("randomstring");
-    await newClient.upload(body, body.length);
-    const result = await newClient.download(0);
-    assert.deepStrictEqual(await bodyToString(result, body.length), body);
-  });
-
-  it("can be created with a url and a credential and an option bag", async () => {
-    const factories = blockBlobClient.pipeline.factories;
-    const credential = factories[factories.length - 1] as SharedKeyCredential;
-    const newClient = new BlockBlobClient(blockBlobClient.url, credential, {
-      retryOptions: {
-        maxTries: 5
-      }
-    });
-
-    const body: string = getUniqueName("randomstring");
-    await newClient.upload(body, body.length);
-    const result = await newClient.download(0);
-    assert.deepStrictEqual(await bodyToString(result, body.length), body);
-  });
-
-  it("can be created with a url and a pipeline", async () => {
-    const factories = blockBlobClient.pipeline.factories;
-    const credential = factories[factories.length - 1] as SharedKeyCredential;
-    const pipeline = newPipeline(credential);
-    const newClient = new BlockBlobClient(blockBlobClient.url, pipeline);
-
-    const body: string = getUniqueName("randomstring");
-    await newClient.upload(body, body.length);
-    const result = await newClient.download(0);
-    assert.deepStrictEqual(await bodyToString(result, body.length), body);
-  });
-
-  it("can be created with a connection string", async () => {
-    const newClient = new BlockBlobClient(
-      getConnectionStringFromEnvironment(),
-      containerName,
-      blobName
-    );
-
-    const body: string = getUniqueName("randomstring");
-    await newClient.upload(body, body.length);
-    const result = await newClient.download(0);
-    assert.deepStrictEqual(await bodyToString(result, body.length), body);
-  });
-
-  it("throws error if constructor containerName parameter is empty", async () => {
-    try {
-      // tslint:disable-next-line: no-unused-expression
-      new BlockBlobClient(getConnectionStringFromEnvironment(), "", "blobName");
-      assert.fail("Expecting an thrown error but didn't get one.");
-    } catch (error) {
-      assert.equal(
-        "Expecting non-empty strings for containerName and blobName parameters",
-        error.message,
-        "Error message is different than expected."
-      );
-    }
-  });
-
-  it("throws error if constructor blobName parameter is empty", async () => {
-    try {
-      // tslint:disable-next-line: no-unused-expression
-      new BlockBlobClient(getConnectionStringFromEnvironment(), "containerName", "");
-      assert.fail("Expecting an thrown error but didn't get one.");
-    } catch (error) {
-      assert.equal(
-        "Expecting non-empty strings for containerName and blobName parameters",
-        error.message,
-        "Error message is different than expected."
-      );
-    }
   });
 });
