+++ sdk/storage/storage-blob/test/node/blobserviceclient.spec.ts
@@ -1,0 +1,64 @@
+import * as assert from "assert";
+
+import * as dotenv from "dotenv";
+import { BlobServiceClient, newPipeline, SharedKeyCredential } from "../../src";
+import { getBSU, getConnectionStringFromEnvironment } from "../utils";
+dotenv.config({ path: "../.env" });
+
+describe("BlobServiceClient", () => {
+  it("can be created with a url and a credential", async () => {
+    const serviceClient = getBSU();
+    const factories = serviceClient.pipeline.factories;
+    const credential = factories[factories.length - 1] as SharedKeyCredential;
+    const newClient = new BlobServiceClient(serviceClient.url, credential);
+
+    const result = await newClient.getProperties();
+
+    assert.ok(typeof result.requestId);
+    assert.ok(result.requestId!.length > 0);
+    assert.ok(typeof result.version);
+    assert.ok(result.version!.length > 0);
+  });
+
+  it("can be created with a url and a credential and an option bag", async () => {
+    const serviceClient = getBSU();
+    const factories = serviceClient.pipeline.factories;
+    const credential = factories[factories.length - 1] as SharedKeyCredential;
+    const newClient = new BlobServiceClient(serviceClient.url, credential, {
+      retryOptions: {
+        maxTries: 5
+      }
+    });
+
+    const result = await newClient.getProperties();
+
+    assert.ok(typeof result.requestId);
+    assert.ok(result.requestId!.length > 0);
+    assert.ok(typeof result.version);
+    assert.ok(result.version!.length > 0);
+  });
+
+  it("can be created with a url and a pipeline", async () => {
+    const serviceClient = getBSU();
+    const factories = serviceClient.pipeline.factories;
+    const credential = factories[factories.length - 1] as SharedKeyCredential;
+    const pipeline = newPipeline(credential);
+    const newClient = new BlobServiceClient(serviceClient.url, pipeline);
+
+    const result = await newClient.getProperties();
+
+    assert.ok(typeof result.requestId);
+    assert.ok(result.requestId!.length > 0);
+    assert.ok(typeof result.version);
+    assert.ok(result.version!.length > 0);
+  });
+
+  it("can be created from a connection string", async () => {
+    const newClient = BlobServiceClient.fromConnectionString(getConnectionStringFromEnvironment());
+
+    const result = await newClient.getProperties();
+
+    assert.ok(typeof result.requestId);
+    assert.ok(result.requestId!.length > 0);
+  });
+});
