+++ sdk/storage/storage-blob/test/pageblobclient.spec.ts
@@ -1,8 +1,6 @@
 import * as assert from "assert";
 import * as dotenv from "dotenv";
-import { newPipeline, SharedKeyCredential } from "../src";
-import { PageBlobClient } from "../src/PageBlobClient";
-import { bodyToString, getBSU, getConnectionStringFromEnvironment, getUniqueName } from "./utils";
+import { bodyToString, getBSU, getUniqueName } from "./utils";
 dotenv.config({ path: "../.env" });
 
 describe("PageBlobClient", () => {
@@ -144,80 +142,5 @@ describe("PageBlobClient", () => {
     await pageBlobClient.updateSequenceNumber("max", 100);
     propertiesResponse = await pageBlobClient.getProperties();
     assert.equal(propertiesResponse.blobSequenceNumber!, 100);
-  });
-
-  it("can be created with a url and a credential", async () => {
-    const factories = pageBlobClient.pipeline.factories;
-    const credential = factories[factories.length - 1] as SharedKeyCredential;
-    const newClient = new PageBlobClient(pageBlobClient.url, credential);
-
-    await newClient.create(512);
-    const result = await newClient.download(0);
-    assert.deepStrictEqual(await bodyToString(result, 512), "\u0000".repeat(512));
-  });
-
-  it("can be created with a url and a credential and an option bag", async () => {
-    const factories = pageBlobClient.pipeline.factories;
-    const credential = factories[factories.length - 1] as SharedKeyCredential;
-    const newClient = new PageBlobClient(pageBlobClient.url, credential, {
-      retryOptions: {
-        maxTries: 5
-      }
-    });
-
-    await newClient.create(512);
-    const result = await newClient.download(0);
-    assert.deepStrictEqual(await bodyToString(result, 512), "\u0000".repeat(512));
-  });
-
-  it("can be created with a url and a pipeline", async () => {
-    const factories = pageBlobClient.pipeline.factories;
-    const credential = factories[factories.length - 1] as SharedKeyCredential;
-    const pipeline = newPipeline(credential);
-    const newClient = new PageBlobClient(pageBlobClient.url, pipeline);
-
-    await newClient.create(512);
-    const result = await newClient.download(0);
-    assert.deepStrictEqual(await bodyToString(result, 512), "\u0000".repeat(512));
-  });
-
-  it("can be created with a connection string", async () => {
-    const newClient = new PageBlobClient(
-      getConnectionStringFromEnvironment(),
-      containerName,
-      blobName
-    );
-
-    await newClient.create(512);
-    const result = await newClient.download(0);
-    assert.deepStrictEqual(await bodyToString(result, 512), "\u0000".repeat(512));
-  });
-
-  it("throws error if constructor containerName parameter is empty", async () => {
-    try {
-      // tslint:disable-next-line: no-unused-expression
-      new PageBlobClient(getConnectionStringFromEnvironment(), "", "blobName");
-      assert.fail("Expecting an thrown error but didn't get one.");
-    } catch (error) {
-      assert.equal(
-        "Expecting non-empty strings for containerName and blobName parameters",
-        error.message,
-        "Error message is different than expected."
-      );
-    }
-  });
-
-  it("throws error if constructor blobName parameter is empty", async () => {
-    try {
-      // tslint:disable-next-line: no-unused-expression
-      new PageBlobClient(getConnectionStringFromEnvironment(), "containerName", "");
-      assert.fail("Expecting an thrown error but didn't get one.");
-    } catch (error) {
-      assert.equal(
-        "Expecting non-empty strings for containerName and blobName parameters",
-        error.message,
-        "Error message is different than expected."
-      );
-    }
   });
 });
