+++ sdk/storage/storage-file/src/FileServiceClient.ts
@@ -111,6 +111,8 @@ export class FileServiceClient extends StorageClient {
   private serviceContext: Service;
 
   /**
+   * ONLY AVAILABLE IN NODE.JS RUNTIME.
+   *
    * Creates an instance of FileServiceClient from connection string.
    *
    * @param {string} connectionString Connection string for an Azure storage account.
