+++ sdk/storage/storage-file/test/directoryclient.spec.ts
@@ -1,7 +1,6 @@
 import * as assert from "assert";
 import { getBSU, getUniqueName } from "./utils";
 import * as dotenv from "dotenv";
-import { DirectoryClient, newPipeline, SharedKeyCredential } from "../src";
 dotenv.config({ path: "../.env" });
 
 describe("DirectoryClient", () => {
@@ -203,52 +202,5 @@ describe("DirectoryClient", () => {
       assert.ok((error.statusCode as number) === 404);
     }
     await subDirClient.delete();
-  });
-
-  it("can be created with a url and a credential", async () => {
-    const factories = dirClient.pipeline.factories;
-    const credential = factories[factories.length - 1] as SharedKeyCredential;
-    const newClient = new DirectoryClient(dirClient.url, credential);
-
-    const result = await newClient.getProperties();
-
-    assert.ok(result.eTag!.length > 0);
-    assert.ok(result.lastModified);
-    assert.ok(result.requestId);
-    assert.ok(result.version);
-    assert.ok(result.date);
-  });
-
-  it("can be created with a url and a credential and an option bag", async () => {
-    const factories = dirClient.pipeline.factories;
-    const credential = factories[factories.length - 1] as SharedKeyCredential;
-    const newClient = new DirectoryClient(dirClient.url, credential, {
-      retryOptions: {
-        maxTries: 5
-      }
-    });
-
-    const result = await newClient.getProperties();
-
-    assert.ok(result.eTag!.length > 0);
-    assert.ok(result.lastModified);
-    assert.ok(result.requestId);
-    assert.ok(result.version);
-    assert.ok(result.date);
-  });
-
-  it("can be created with a url and a pipeline", async () => {
-    const factories = dirClient.pipeline.factories;
-    const credential = factories[factories.length - 1] as SharedKeyCredential;
-    const pipeline = newPipeline(credential);
-    const newClient = new DirectoryClient(dirClient.url, pipeline);
-
-    const result = await newClient.getProperties();
-
-    assert.ok(result.eTag!.length > 0);
-    assert.ok(result.lastModified);
-    assert.ok(result.requestId);
-    assert.ok(result.version);
-    assert.ok(result.date);
   });
 });
