+++ sdk/storage/storage-file/test/node/fileclient.spec.ts
@@ -2,6 +2,7 @@ import * as assert from "assert";
 import { Duplex } from "stream";
 import { bodyToString, getBSU, getUniqueName } from "../utils";
 import { Buffer } from "buffer";
+import { FileClient, newPipeline, SharedKeyCredential } from "../../src";
 
 describe("FileClient Node.js only", () => {
   const serviceClient = getBSU();
@@ -11,6 +12,7 @@ describe("FileClient Node.js only", () => {
   let dirClient = shareClient.createDirectoryClient(dirName);
   let fileName = getUniqueName("file");
   let fileClient = dirClient.createFileClient(fileName);
+  const content = "Hello World";
 
   beforeEach(async () => {
     shareName = getUniqueName("share");
@@ -65,5 +67,73 @@ describe("FileClient Node.js only", () => {
     await fileClient.uploadRange(body, 0, bodyLength);
     const result = await fileClient.download(0);
     assert.deepStrictEqual(await bodyToString(result, bodyLength), body);
+  });
+
+  it("can be created with a url and a credential", async () => {
+    await fileClient.create(content.length);
+    const metadata = {
+      a: "a",
+      b: "b"
+    };
+    await fileClient.setMetadata(metadata);
+
+    const factories = fileClient.pipeline.factories;
+    const credential = factories[factories.length - 1] as SharedKeyCredential;
+    const newClient = new FileClient(fileClient.url, credential);
+
+    const result = await newClient.getProperties();
+
+    assert.ok(result.eTag!.length > 0);
+    assert.ok(result.lastModified);
+    assert.ok(result.requestId);
+    assert.ok(result.version);
+    assert.ok(result.date);
+  });
+
+  it("can be created with a url and a credential and an option bag", async () => {
+    await fileClient.create(content.length);
+    const metadata = {
+      a: "a",
+      b: "b"
+    };
+    await fileClient.setMetadata(metadata);
+
+    const factories = fileClient.pipeline.factories;
+    const credential = factories[factories.length - 1] as SharedKeyCredential;
+    const newClient = new FileClient(fileClient.url, credential, {
+      retryOptions: {
+        maxTries: 5
+      }
+    });
+
+    const result = await newClient.getProperties();
+
+    assert.ok(result.eTag!.length > 0);
+    assert.ok(result.lastModified);
+    assert.ok(result.requestId);
+    assert.ok(result.version);
+    assert.ok(result.date);
+  });
+
+  it("can be created with a url and a pipeline", async () => {
+    await fileClient.create(content.length);
+    const metadata = {
+      a: "a",
+      b: "b"
+    };
+    await fileClient.setMetadata(metadata);
+
+    const factories = fileClient.pipeline.factories;
+    const credential = factories[factories.length - 1] as SharedKeyCredential;
+    const pipeline = newPipeline(credential);
+    const newClient = new FileClient(fileClient.url, pipeline);
+
+    const result = await newClient.getProperties();
+
+    assert.ok(result.eTag!.length > 0);
+    assert.ok(result.lastModified);
+    assert.ok(result.requestId);
+    assert.ok(result.version);
+    assert.ok(result.date);
   });
 });
