+++ sdk/storage/storage-file/test/node/fileserviceclient.spec.ts
@@ -1,0 +1,74 @@
+import * as assert from "assert";
+import { getBSU, getConnectionStringFromEnvironment } from "../utils";
+import * as dotenv from "dotenv";
+import { FileServiceClient, newPipeline, SharedKeyCredential } from "../../src";
+dotenv.config({ path: "../.env" });
+
+describe("FileServiceClient", () => {
+  it("can be created with a url and a credential", async () => {
+    const serviceClient = getBSU();
+    const factories = serviceClient.pipeline.factories;
+    const credential = factories[factories.length - 1] as SharedKeyCredential;
+    const newClient = new FileServiceClient(serviceClient.url, credential);
+
+    const result = await newClient.getProperties();
+
+    assert.ok(typeof result.requestId);
+    assert.ok(result.requestId!.length > 0);
+    assert.ok(typeof result.version);
+    assert.ok(result.version!.length > 0);
+  });
+
+  it("can be created with a url and a credential and an option bag", async () => {
+    const serviceClient = getBSU();
+    const factories = serviceClient.pipeline.factories;
+    const credential = factories[factories.length - 1] as SharedKeyCredential;
+    const newClient = new FileServiceClient(serviceClient.url, credential, {
+      retryOptions: { maxTries: 5 }
+    });
+
+    const result = await newClient.getProperties();
+
+    assert.ok(typeof result.requestId);
+    assert.ok(result.requestId!.length > 0);
+    assert.ok(typeof result.version);
+    assert.ok(result.version!.length > 0);
+  });
+
+  it("can be created with a url and a pipeline", async () => {
+    const serviceClient = getBSU();
+    const factories = serviceClient.pipeline.factories;
+    const credential = factories[factories.length - 1] as SharedKeyCredential;
+    const pipeline = newPipeline(credential);
+    const newClient = new FileServiceClient(serviceClient.url, pipeline);
+
+    const result = await newClient.getProperties();
+
+    assert.ok(typeof result.requestId);
+    assert.ok(result.requestId!.length > 0);
+    assert.ok(typeof result.version);
+    assert.ok(result.version!.length > 0);
+  });
+
+  it("can be created from a connection string", async () => {
+    const newClient = FileServiceClient.fromConnectionString(getConnectionStringFromEnvironment());
+
+    const result = await newClient.getProperties();
+
+    assert.ok(typeof result.requestId);
+    assert.ok(result.requestId!.length > 0);
+  });
+
+  it("can be created from a connection string and an option bag", async () => {
+    const newClient = FileServiceClient.fromConnectionString(getConnectionStringFromEnvironment(), {
+      retryOptions: {
+        maxTries: 5
+      }
+    });
+
+    const result = await newClient.getProperties();
+
+    assert.ok(typeof result.requestId);
+    assert.ok(result.requestId!.length > 0);
+  });
+});
