+++ sdk/storage/storage-file/test/node/shareclient.spec.ts
@@ -1,6 +1,6 @@
 import * as assert from "assert";
-import { SignedIdentifier } from "../../src/ShareClient";
-import { getBSU, getUniqueName } from "./../utils";
+import { newPipeline, ShareClient, SharedKeyCredential, SignedIdentifier } from "../../src";
+import { getBSU, getUniqueName, getConnectionStringFromEnvironment } from "./../utils";
 
 describe("ShareClient", () => {
   const serviceClient = getBSU();
@@ -55,5 +55,92 @@ describe("ShareClient", () => {
   it("getAccessPolicy", (done) => {
     // create() with default parameters has been tested in setAccessPolicy
     done();
+  });
+
+  it("can be created with a url and a credential", async () => {
+    const factories = shareClient.pipeline.factories;
+    const credential = factories[factories.length - 1] as SharedKeyCredential;
+    const newClient = new ShareClient(shareClient.url, credential);
+
+    const result = await newClient.getProperties();
+
+    assert.ok(result.eTag!.length > 0);
+    assert.ok(result.lastModified);
+    assert.ok(result.requestId);
+    assert.ok(result.version);
+    assert.ok(result.date);
+  });
+
+  it("can be created with a url and a credential and an option bag", async () => {
+    const factories = shareClient.pipeline.factories;
+    const credential = factories[factories.length - 1] as SharedKeyCredential;
+    const newClient = new ShareClient(shareClient.url, credential, {
+      retryOptions: {
+        maxTries: 5
+      }
+    });
+
+    const result = await newClient.getProperties();
+
+    assert.ok(result.eTag!.length > 0);
+    assert.ok(result.lastModified);
+    assert.ok(result.requestId);
+    assert.ok(result.version);
+    assert.ok(result.date);
+  });
+
+  it("can be created with a url and a pipeline", async () => {
+    const factories = shareClient.pipeline.factories;
+    const credential = factories[factories.length - 1] as SharedKeyCredential;
+    const pipeline = newPipeline(credential);
+    const newClient = new ShareClient(shareClient.url, pipeline);
+
+    const result = await newClient.getProperties();
+
+    assert.ok(result.eTag!.length > 0);
+    assert.ok(result.lastModified);
+    assert.ok(result.requestId);
+    assert.ok(result.version);
+    assert.ok(result.date);
+  });
+
+  it("can be created with a connection string and a share name", async () => {
+    const newClient = new ShareClient(getConnectionStringFromEnvironment(), shareName);
+    const result = await newClient.getProperties();
+
+    assert.ok(result.eTag!.length > 0);
+    assert.ok(result.lastModified);
+    assert.ok(result.requestId);
+    assert.ok(result.version);
+    assert.ok(result.date);
+  });
+
+  it("can be created with a connection string and a share name and an option bag", async () => {
+    const newClient = new ShareClient(getConnectionStringFromEnvironment(), shareName, {
+      retryOptions: {
+        maxTries: 5
+      }
+    });
+    const result = await newClient.getProperties();
+
+    assert.ok(result.eTag!.length > 0);
+    assert.ok(result.lastModified);
+    assert.ok(result.requestId);
+    assert.ok(result.version);
+    assert.ok(result.date);
+  });
+
+  it("throws error if constructor shareName parameter is empty", async () => {
+    try {
+      // tslint:disable-next-line: no-unused-expression
+      new ShareClient(getConnectionStringFromEnvironment(), "");
+      assert.fail("Expecting an thrown error but didn't get one.");
+    } catch (error) {
+      assert.equal(
+        "Expecting non-empty strings for shareName parameter",
+        error.message,
+        "Error message is different than expected."
+      );
+    }
   });
 });
