+++ sdk/storage/storage-file/test/shareclient.spec.ts
@@ -1,7 +1,6 @@
 import * as assert from "assert";
 import * as dotenv from "dotenv";
-import { newPipeline, ShareClient, SharedKeyCredential } from "../src";
-import { getBSU, getConnectionStringFromEnvironment, getUniqueName } from "./utils";
+import { getBSU, getUniqueName } from "./utils";
 dotenv.config({ path: "../.env" });
 
 describe("ShareClient", () => {
@@ -130,92 +129,5 @@ describe("ShareClient", () => {
     const root = await shareClient.rootDirectoryClient;
     const result = await root.getProperties();
     assert.ok(result, "Expecting valid properties for the root directory.");
-  });
-
-  it("can be created with a url and a credential", async () => {
-    const factories = shareClient.pipeline.factories;
-    const credential = factories[factories.length - 1] as SharedKeyCredential;
-    const newClient = new ShareClient(shareClient.url, credential);
-
-    const result = await newClient.getProperties();
-
-    assert.ok(result.eTag!.length > 0);
-    assert.ok(result.lastModified);
-    assert.ok(result.requestId);
-    assert.ok(result.version);
-    assert.ok(result.date);
-  });
-
-  it("can be created with a url and a credential and an option bag", async () => {
-    const factories = shareClient.pipeline.factories;
-    const credential = factories[factories.length - 1] as SharedKeyCredential;
-    const newClient = new ShareClient(shareClient.url, credential, {
-      retryOptions: {
-        maxTries: 5
-      }
-    });
-
-    const result = await newClient.getProperties();
-
-    assert.ok(result.eTag!.length > 0);
-    assert.ok(result.lastModified);
-    assert.ok(result.requestId);
-    assert.ok(result.version);
-    assert.ok(result.date);
-  });
-
-  it("can be created with a url and a pipeline", async () => {
-    const factories = shareClient.pipeline.factories;
-    const credential = factories[factories.length - 1] as SharedKeyCredential;
-    const pipeline = newPipeline(credential);
-    const newClient = new ShareClient(shareClient.url, pipeline);
-
-    const result = await newClient.getProperties();
-
-    assert.ok(result.eTag!.length > 0);
-    assert.ok(result.lastModified);
-    assert.ok(result.requestId);
-    assert.ok(result.version);
-    assert.ok(result.date);
-  });
-
-  it("can be created with a connection string and a share name", async () => {
-    const newClient = new ShareClient(getConnectionStringFromEnvironment(), shareName);
-    const result = await newClient.getProperties();
-
-    assert.ok(result.eTag!.length > 0);
-    assert.ok(result.lastModified);
-    assert.ok(result.requestId);
-    assert.ok(result.version);
-    assert.ok(result.date);
-  });
-
-  it("can be created with a connection string and a share name and an option bag", async () => {
-    const newClient = new ShareClient(getConnectionStringFromEnvironment(), shareName, {
-      retryOptions: {
-        maxTries: 5
-      }
-    });
-    const result = await newClient.getProperties();
-
-    assert.ok(result.eTag!.length > 0);
-    assert.ok(result.lastModified);
-    assert.ok(result.requestId);
-    assert.ok(result.version);
-    assert.ok(result.date);
-  });
-
-  it("throws error if constructor shareName parameter is empty", async () => {
-    try {
-      // tslint:disable-next-line: no-unused-expression
-      new ShareClient(getConnectionStringFromEnvironment(), "");
-      assert.fail("Expecting an thrown error but didn't get one.");
-    } catch (error) {
-      assert.equal(
-        "Expecting non-empty strings for shareName parameter",
-        error.message,
-        "Error message is different than expected."
-      );
-    }
   });
 });
