+++ sdk/storage/storage-queue/test/messageidclient.spec.ts
@@ -3,7 +3,6 @@ import { getQSU, sleep } from "./utils";
 import { QueueClient } from "../src/QueueClient";
 import { record } from "./utils/recorder";
 import * as dotenv from "dotenv";
-import { SharedKeyCredential, MessageIdClient, newPipeline } from "../src";
 dotenv.config({ path: "../.env" });
 
 describe("MessageIdClient", () => {
@@ -160,58 +159,5 @@ describe("MessageIdClient", () => {
       error = err;
     }
     assert.ok(error);
-  });
-
-  it("can be created with a url and a credential", async () => {
-    const messagesClient = queueClient.createMessagesClient();
-
-    const eResult = await messagesClient.enqueue(messageContent);
-    assert.ok(eResult.date);
-
-    const newMessage = "";
-    const messageIdClient = messagesClient.createMessageIdClient(eResult.messageId);
-    const factories = messagesClient.pipeline.factories;
-    const credential = factories[factories.length - 1] as SharedKeyCredential;
-    const newClient = new MessageIdClient(messageIdClient.url, credential);
-    const uResult = await newClient.update(eResult.popReceipt, newMessage);
-
-    assert.ok(uResult.version);
-  });
-
-  it("can be created with a url and a credential and an option bag", async () => {
-    const messagesClient = queueClient.createMessagesClient();
-
-    const eResult = await messagesClient.enqueue(messageContent);
-    assert.ok(eResult.date);
-
-    const newMessage = "";
-    const messageIdClient = messagesClient.createMessageIdClient(eResult.messageId);
-    const factories = messagesClient.pipeline.factories;
-    const credential = factories[factories.length - 1] as SharedKeyCredential;
-    const newClient = new MessageIdClient(messageIdClient.url, credential, {
-      retryOptions: {
-        maxTries: 5
-      }
-    });
-    const uResult = await newClient.update(eResult.popReceipt, newMessage);
-
-    assert.ok(uResult.version);
-  });
-
-  it("can be created with a url and a pipeline", async () => {
-    const messagesClient = queueClient.createMessagesClient();
-
-    const eResult = await messagesClient.enqueue(messageContent);
-    assert.ok(eResult.date);
-
-    const newMessage = "";
-    const messageIdClient = messagesClient.createMessageIdClient(eResult.messageId);
-    const factories = messagesClient.pipeline.factories;
-    const credential = factories[factories.length - 1] as SharedKeyCredential;
-    const pipeline = newPipeline(credential);
-    const newClient = new MessageIdClient(messageIdClient.url, pipeline);
-    const uResult = await newClient.update(eResult.popReceipt, newMessage);
-
-    assert.ok(uResult.version);
   });
 });
